examples/python/requirements.txt
snowflake-connector-python>=3.13,<4
python-dotenv>=1.0,<2
